Moved all admin routes to admin/api.
