+++ lib/core/routes.js
@@ -78,18 +78,18 @@ function routes(app) {
 	// Cloudinary API for selecting an existing image / upload a new image
 	if (keystone.get('cloudinary config')) {
 		debug('setting cloudinary api');
-		app.get('/keystone/api/cloudinary/get', require('../../routes/api/cloudinary').get);
-		app.get('/keystone/api/cloudinary/autocomplete', require('../../routes/api/cloudinary').autocomplete);
-		app.post('/keystone/api/cloudinary/upload', require('../../routes/api/cloudinary').upload);
+		app.get('/keystone/api/cloudinary/get', require('../../admin/api/cloudinary').get);
+		app.get('/keystone/api/cloudinary/autocomplete', require('../../admin/api/cloudinary').autocomplete);
+		app.post('/keystone/api/cloudinary/upload', require('../../admin/api/cloudinary').upload);
 	}
 
 	// Generic Lists API
 	debug('setting generic list api');
-	app.all('/keystone/api/:list/:action(autocomplete|get|order|create|delete|fetch)', initList(), require('../../routes/api/list'));
+	app.all('/keystone/api/:list/:action(autocomplete|get|order|create|delete|fetch)', initList(), require('../../admin/api/list'));
 	app.get('/keystone/api/:list/:id', initList(), require('../../admin/api/item/get'));
 
 	debug('setting generic Lists download route');
-	app.all('/keystone/download/:list', initList(), require('../../routes/download/list'));
+	app.all('/keystone/download/:list', initList(), require('../../admin/api/download'));
 
 	// Admin-UI API
 	debug('setting list and item details admin routes');
